pyproject.toml
[project]
name = "ai-researcher"
version = "0.1.0"
description = "AI Academic Research Agent for Comparative Literature"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "litellm>=1.50.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.38.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "respx>=0.21.0",
]

[project.scripts]
ai-researcher = "cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta:__legacy__"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit real external APIs (may be slow)",
]
